Serializer: Fixed to_pod for subclasses of PODable types
We were only checking if the values type  matched a type in the
`POD_TYPES` list, what we should have been checking is if it is
an instance of a type in `POD_TYPES`.

This commit also adds NoneType to the list as this is also a valid
PODable type.